# OpenAPI.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /products:
    get:
      tags:
      - product-controller
      operationId: getAllProducts
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CursorPageResponseProductDTO"
    post:
      tags:
      - product-controller
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /order:
    get:
      tags:
      - order-controller
      operationId: getAllOrders
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CursorPageResponseOrderDTO"
    post:
      tags:
      - order-controller
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /order/{orderId}/products/{productId}:
    post:
      tags:
      - order-controller
      operationId: addProductToOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /customer:
    get:
      tags:
      - customer-controller
      operationId: getAllCustomers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerDTO"
    post:
      tags:
      - customer-controller
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CustomerDTO"
  /products/{id}:
    get:
      tags:
      - product-controller
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /order/{id}:
    get:
      tags:
      - order-controller
      operationId: getOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
components:
  schemas:
    ProductRequest:
      type: object
      properties:
        description:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orderIds:
          type: array
          items:
            type: integer
            format: int64
    OrderRequest:
      type: object
      properties:
        description:
          type: string
        customerId:
          type: integer
          format: int64
        productIds:
          type: array
          items:
            type: integer
            format: int64
    OrderCustomerDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          $ref: "#/components/schemas/OrderCustomerDTO"
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductOrderDTO"
    ProductOrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
    CustomerDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/CustomerOrderDTO"
    CustomerOrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
    CursorPageResponseProductDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        pageSize:
          type: integer
          format: int32
        nextCursor:
          type: integer
          format: int64
        hasNext:
          type: boolean
    CursorPageResponseOrderDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrderDTO"
        pageSize:
          type: integer
          format: int32
        nextCursor:
          type: integer
          format: int64
        hasNext:
          type: boolean